+++ src/main/java/org/liquidengine/legui/component/misc/listener/scrollbar/ScrollBarScrollListener.java
@@ -4,7 +4,6 @@ import org.liquidengine.legui.component.ScrollBar;
 import org.liquidengine.legui.component.event.scrollbar.ScrollBarChangeValueEvent;
 import org.liquidengine.legui.event.ScrollEvent;
 import org.liquidengine.legui.listener.ScrollEventListener;
-import org.liquidengine.legui.listener.processor.EventProcessor;
 
 import static org.liquidengine.legui.component.misc.listener.scrollbar.ScrollBarHelper.updateScrollBarValue;
